requirements.txt
matplotlib==3.2.2
scipy==1.4.1
Werkzeug==1.0.1
seaborn==0.10.1
numpy==1.19.0
twilio==6.44.1
Flask==1.1.2
pydub==0.24.1
PyAudio==0.2.11
pandas==1.0.5
protobuf==3.12.4
python-dotenv==0.14.0
scikit_learn==0.23.1
